added scanning fragment
