+++ .idea/misc.xml
@@ -6,6 +6,7 @@
         <entry key="..\:/Users/rajan/AndroidStudioProjects/inQuiRy/app/src/main/res/layout/activity_main.xml" value="0.335" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.41309255079006774" />
         <entry key="app/src/main/res/layout/activity_scanner.xml" value="0.3658155289213132" />
+        <entry key="app/src/main/res/layout/scanner_fragment.xml" value="0.4136933461909354" />
       </map>
     </option>
   </component>
