+++ app/src/main/java/com/cmput301w22t13/inquiry/classes/Scanner.java
@@ -26,144 +26,4 @@ import org.jetbrains.annotations.NotNull;
 public class Scanner extends AppCompatActivity {
 
 
-    private CodeScanner mCodeScanner;
-    boolean CameraPermission = false;
-    final int CAMERA_PERM = 1;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        CodeScannerView scannerView = findViewById(R.id.scanner_view);
-        mCodeScanner = new CodeScanner(this,scannerView);
-        askPermission();
-        if (CameraPermission) {
-
-            scannerView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-
-                    mCodeScanner.startPreview();
-
-                }
-            });
-
-            mCodeScanner.setDecodeCallback(new DecodeCallback() {
-                @Override
-                public void onDecoded(@NonNull com.google.zxing.Result result) {
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-
-                            Toast.makeText(Scanner.this,
-                                    result.getText(),
-                                    Toast.LENGTH_LONG).show();
-                        }
-                    });
-                }
-            });
-
-        }
-    }
-
-    private void askPermission(){
-
-        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP){
-
-
-            if (ActivityCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED ){
-
-                ActivityCompat.requestPermissions(Scanner.this,new String[]{Manifest.permission.CAMERA},CAMERA_PERM);
-
-
-            }else {
-
-                mCodeScanner.startPreview();
-                CameraPermission = true;
-            }
-
-        }
-
-
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull @NotNull String[] permissions, @NonNull @NotNull int[] grantResults) {
-
-        if (requestCode == CAMERA_PERM){
-
-
-            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED){
-
-                mCodeScanner.startPreview();
-                CameraPermission = true;
-            }else {
-
-                if (ActivityCompat.shouldShowRequestPermissionRationale(this,Manifest.permission.CAMERA)){
-
-                    new AlertDialog.Builder(this)
-                            .setTitle("Permission")
-                            .setMessage("Please provide the camera permission for using all the features of the app")
-                            .setPositiveButton("Proceed", new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface dialog, int which) {
-
-                                    ActivityCompat.requestPermissions(Scanner.this,new String[]{Manifest.permission.CAMERA},CAMERA_PERM);
-
-                                }
-                            }).setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-
-                            dialog.dismiss();
-
-                        }
-                    }).create().show();
-
-                }else {
-
-                    new AlertDialog.Builder(this)
-                            .setTitle("Permission")
-                            .setMessage("You have denied some permission. Allow all permission at [Settings] > [Permissions]")
-                            .setPositiveButton("Settings", new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface dialog, int which) {
-
-                                    dialog.dismiss();
-                                    Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS,
-                                            Uri.fromParts("package",getPackageName(),null));
-                                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                                    startActivity(intent);
-                                    finish();
-
-
-                                }
-                            }).setNegativeButton("No, Exit app", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-
-                            dialog.dismiss();
-                            finish();
-
-                        }
-                    }).create().show();
-
-
-
-                }
-
-            }
-
-        }
-
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-    }
-
-    @Override
-    protected void onPause() {
-        if (CameraPermission){
-            mCodeScanner.releaseResources();
-        }
-
-        super.onPause();
-    }
 }
