+++ app/src/main/java/com/cmput301w22t13/inquiry/ui/scanner/ScannerFragment.java
@@ -3,6 +3,7 @@ package com.cmput301w22t13.inquiry.ui.scanner;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
+import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 
@@ -14,36 +15,56 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import android.widget.Toast;
 
+import com.budiyev.android.codescanner.CodeScanner;
+import com.budiyev.android.codescanner.CodeScannerView;
+import com.budiyev.android.codescanner.DecodeCallback;
+import com.cmput301w22t13.inquiry.R;
 import com.cmput301w22t13.inquiry.activities.ScanActivity;
 import com.cmput301w22t13.inquiry.databinding.ScannerFragmentBinding;
+import com.google.zxing.Result;
 
 public class ScannerFragment extends Fragment {
 
-    private ScannerViewModel scannerViewModel;
-    private ScannerFragmentBinding binding;
+    private CodeScanner mCodeScanner;
 
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        scannerViewModel = new ViewModelProvider(this).get(ScannerViewModel.class);
-        binding = ScannerFragmentBinding.inflate(inflater, container, false);
-        View root = binding.getRoot();
-
-        final TextView textView = binding.textScanner;
-        scannerViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        final Activity activity = getActivity();
+        View root = inflater.inflate(R.layout.scanner_fragment, container, false);
+        CodeScannerView scannerView = root.findViewById(R.id.scanner_view);
+        mCodeScanner = new CodeScanner(activity, scannerView);
+        mCodeScanner.setDecodeCallback(new DecodeCallback() {
             @Override
-            public void onChanged(@Nullable String s) {
-                textView.setText(s);
+            public void onDecoded(@NonNull final Result result) {
+                activity.runOnUiThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        Toast.makeText(activity, result.getText(), Toast.LENGTH_SHORT).show();
+                    }
+                });
             }
         });
-        Intent intent = new Intent(getActivity(), ScanActivity.class);
-        startActivity(intent);
+        scannerView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                mCodeScanner.startPreview();
+            }
+        });
         return root;
     }
 
     @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        binding = null;
+    public void onResume() {
+        super.onResume();
+        mCodeScanner.startPreview();
     }
-}
+
+    @Override
+    public void onPause() {
+        mCodeScanner.releaseResources();
+        super.onPause();
+    }}
